Move Resolver out of __init__.py
